Move function to split audio to slixer command utils file
